refactor(salarié): Déplace le composant Cotisation dans son propre fichier
