+++ site/source/components/PaySlipCotisation.tsx
@@ -1,0 +1,65 @@
+import { DottedName } from 'modele-social'
+import { ASTNode, formatValue, reduceAST } from 'publicodes'
+import { useContext } from 'react'
+import { useTranslation } from 'react-i18next'
+
+import RuleLink from '@/components/RuleLink'
+import { EngineContext } from '@/components/utils/EngineContext'
+
+export default function Cotisation({ dottedName }: { dottedName: DottedName }) {
+	const language = useTranslation().i18n.language
+	const engine = useContext(EngineContext)
+	const partSalariale = engine.evaluate(
+		findReferenceInNode(
+			dottedName,
+			engine.getRule('salarié . cotisations . salarié')
+		) ?? '0'
+	)
+	const partPatronale = engine.evaluate(
+		findReferenceInNode(
+			dottedName,
+			engine.getRule('salarié . cotisations . employeur')
+		) ?? '0'
+	)
+
+	if (!partPatronale.nodeValue && !partSalariale.nodeValue) {
+		return null
+	}
+
+	return (
+		<>
+			<RuleLink dottedName={dottedName} />
+			<span>
+				{partPatronale?.nodeValue
+					? formatValue(partPatronale, { displayedUnit: '€', language })
+					: '–'}
+			</span>
+			<span>
+				{partSalariale?.nodeValue
+					? formatValue(partSalariale, { displayedUnit: '€', language })
+					: '–'}
+			</span>
+		</>
+	)
+}
+
+function findReferenceInNode(
+	dottedName: DottedName,
+	node: ASTNode
+): string | undefined {
+	return reduceAST<string | undefined>(
+		(acc, node) => {
+			if (
+				node.nodeKind === 'reference' &&
+				node.dottedName?.startsWith(dottedName) &&
+				!node.dottedName.endsWith('$SITUATION')
+			) {
+				return node.dottedName
+			} else if (node.nodeKind === 'reference') {
+				return acc
+			}
+		},
+		undefined,
+		node
+	)
+}
